Decouple builder from framework image
Currently the framework image is used both for building
the binary and as the actual artifact.
Eventually we end up with large framework image that includes
other packages it doesn't need.

Building the framework binary is performed by using alpine-go
image directly.
The framework image is built from alpine image and includes
only the binary.
In order to reduce build time and provide better development experience,
go runtime cache is saved to a persistent volume, that is being reused
on each iteration.

Also update the framework image name.

Signed-off-by: Or Mergi <ormergi@redhat.com>